requirements.txt
fastapi
uvicorn
google-generativeai
sentence-transformers
faiss-cpu
youtube-transcript-api
python-dotenv
